+++ src/lib/tokenizer/bpe_nonpar_core.cpp
@@ -1,9 +1,19 @@
 // bpe_loop_core.h
 
 #pragma once
-#include "bpe_loop_core.h"
+#include "bpe_core.h"
+
+// 匿名的命名空间, 等价于声明 静态存储 & 本文件私有
+namespace {
+    // thread_local 声明线程隔离变量
+
+}
 
 
+
+
+
+extern "C" {
 
 std::vector<std::pair<std::pair<uint32_t, uint32_t>, uint64_t>> c_nonpar_bpe(
     const int num_merges,
@@ -11,4 +21,8 @@ std::vector<std::pair<std::pair<uint32_t, uint32_t>, uint64_t>> c_nonpar_bpe(
     const uint32_t* tokens_ptr,
     const int64_t* offsets_ptr,
     const uint64_t* freqs_ptr
-);
+) {
+    
+}
+
+} // end extern "C"
